0.71
THIS DAY I PUT AN END TO THE STANGE BUG THAT RUINS THE GAME APON RESTARTING.

So i was subscribing to an event on the Awake() function. However for some reason that means it is STILL SUBSCRIBED even after reloading the scene. So now the event is subscribed to onEnable(), and unsubbed in OnDisable().